requirements.txt
streamlit>=1.39.0
requests>=2.32.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
numpy>=1.26.0
